requirements.txt
absl-py==2.0.0
alabaster==0.7.16
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
appnope==0.1.3
astatine==0.3.3
astor==0.8.1
astpretty==3.0.0
astroid==2.15.8
asttokens==2.0.8
attrs==23.2.0
autoflake==1.7.8
Babel==2.14.0
backcall==0.2.0
bandit==1.7.6
beautifulsoup4==4.12.2
black==23.12.1
bleach==5.0.1
bokeh==2.4.3
cachetools==5.3.2
certifi==2022.9.24
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.28.1
cognitive_complexity==1.3.0
colorama==0.4.6
comm==0.2.1
contourpy==1.0.5
coverage==7.4.0
cycler==0.11.0
darglint==1.8.1
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.8
dlint==0.14.1
doc8==1.1.1
docformatter==1.7.5
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.19
domdf-python-tools==3.8.0.post2
entrypoints==0.4
eradicate==2.3.0
executing==1.1.1
fastapi==0.109.0
fastjsonschema==2.19.1
filelock==3.13.1
flake8==5.0.4
flake8-2020==1.8.1
flake8-aaa==0.17.0
flake8-annotations==3.0.1
flake8-annotations-complexity==0.0.8
flake8-annotations-coverage==0.0.6
flake8-bandit==4.1.1
flake8-black==0.3.6
flake8-blind-except==0.2.1
flake8-breakpoint==1.1.0
flake8-broken-line==0.6.0
flake8-bugbear==23.3.12
flake8-builtins==1.5.3
flake8-class-attributes-order==0.1.3
flake8-coding==1.3.2
flake8-comments==0.1.2
flake8-comprehensions==3.14.0
flake8-debugger==4.1.2
flake8-django==1.4
flake8-docstrings==1.7.0
flake8-encodings==0.5.1
flake8-eradicate==1.5.0
flake8-executable==2.1.3
flake8-expression-complexity==0.0.11
flake8-fastapi==0.7.0
flake8-fixme==1.1.1
flake8-functions==0.0.8
flake8-future-annotations==0.0.5
flake8-helper==0.2.2
flake8-isort==6.1.1
flake8-literal==1.4.0
flake8-logging-format==0.9.0
flake8-markdown==0.5.0
flake8-mutable==1.2.0
flake8-no-pep420==2.7.0
flake8-noqa==1.4.0
flake8-pie==0.16.0
flake8-plugin-utils==1.3.3
flake8-pyi==22.11.0
flake8-pylint==0.2.1
flake8-pytest-style==1.7.2
flake8-quotes==3.3.2
flake8-rst-docstrings==0.3.0
flake8-secure-coding-standard==1.4.1
flake8-string-format==0.3.0
flake8-tidy-imports==4.10.0
flake8-typing-imports==1.15.0
flake8-use-fstring==1.4
flake8-use-pathlib==0.3.0
flake8-useless-assert==0.4.4
flake8-variables-names==0.0.6
flake8-warnings==0.4.0
flake8_cognitive_complexity==0.1.0
flake8_functions_names==0.4.0
flake8_simplify==0.21.0
fonttools==4.37.4
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.26.2
google-auth-oauthlib==1.2.0
grpcio==1.60.0
gudhi==3.9.0
hypothesis==6.93.1
hypothesmith==0.1.9
idna==3.4
imageio==2.22.2
imagesize==1.4.1
importlib-metadata==5.0.0
iniconfig==2.0.0
ipykernel==6.16.0
ipython==8.5.0
ipywidgets==8.0.2
isort==5.13.2
jaxtyping==0.2.25
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter_bokeh==3.0.5
jupyter_client==7.4.2
jupyter_core==5.7.1
jupyterlab-flake8==0.7.1
jupyterlab-widgets==3.0.3
jupyterlab_pygments==0.3.0
kiwisolver==1.4.4
lark-parser==0.12.0
lazy-object-proxy==1.10.0
lazy_loader==0.3
libcst==0.4.10
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
monai==1.0.0
mpmath==1.3.0
mr-proper==0.0.7
mypy==0.982
mypy-extensions==0.4.3
natsort==8.4.0
nbclient==0.9.0
nbconvert==7.14.1
nbformat==5.9.2
nest-asyncio==1.5.6
networkx==2.8.7
nibabel==5.2.0
numpy==1.23.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
packaging==23.2
pandas==1.5.0
pandas-vet==0.2.3
pandocfilters==1.5.0
panel==0.14.0
param==1.12.2
parso==0.8.3
pathspec==0.12.1
pbr==6.0.0
pep8-naming==0.13.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.4.11
platformdirs==4.1.0
pluggy==1.3.0
POT==0.9.3
prompt-toolkit==3.0.31
protobuf==4.23.4
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybetter==0.4.1
pybind11==2.11.1
pybind11-global==2.11.1
pycln==2.4.0
pycodestyle==2.9.1
pyct==0.4.8
pydantic==2.5.3
pydantic_core==2.14.6
pydocstyle==6.3.0
pyemojify==0.2.0
pyflakes==2.5.0
Pygments==2.13.0
pylint==2.17.7
pyparsing==3.0.9
pyproject-api==1.6.1
pytest==7.4.4
pytest-cov==4.1.0
pytest-sugar==0.9.7
python-dateutil==2.8.2
python-dev-tools==2023.3.24
pytz==2022.4
pyupgrade==3.15.0
pyviz-comms==2.2.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
referencing==0.32.1
removestar==1.5
requests==2.31.0
requests-oauthlib==1.3.1
restructuredtext_lint==1.4.0
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.3.2
scipy==1.9.2
sentry-sdk==1.39.2
setproctitle==1.3.3
six==1.16.0
sklearn==0.0
smmap==5.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==6.2.1
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
ssort==0.12.0
stack-data==0.5.1
starlette==0.35.1
stdlib-list==0.10.0
stevedore==5.1.0
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.1.0
tifffile==2022.10.10
tinycss2==1.2.1
tokenize-rt==5.2.0
tomli==2.0.1
tomlkit==0.12.3
torch==2.1.2
torch-tb-profiler==0.4.3
torchaudio==2.1.2
torchsummary==1.5.1
torchvision==0.16.2
tornado==6.2
tox==4.12.0
tox-travis==0.12
tqdm==4.64.1
traitlets==5.4.0
triton==2.1.0
typeguard==2.13.3
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
untokenize==0.1.1
urllib3==1.26.12
virtualenv==20.25.0
wandb==0.16.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==3.0.1
widgetsnbextension==4.0.3
wrapt==1.16.0
yarg==0.1.9
zipp==3.9.0
